+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Added a `konverter-vault show` command that supports YAML and JSON output
+
 ## [v0.2.0] - 2020-03-01
 
 - Added support for directories as context path